+++ CHANGES.rst
@@ -6,6 +6,7 @@ CHANGELOG for LaunchKey Python SDK
 
 * Added get_all_directory_sdk_keys method to Organization client
 * Added integration testing suite
+* Added device failure sensor type
 
 3.4.0
 -----